+++ src/utils/classes/GenericFormat.ts
