+++ src/utils/classes/KSGeneralTuteeFormat.ts
